requirements.txt
Click>=6.0
PyYAML>=3.12
Jinja2>=2.10
